# openapi.yaml
openapi: 3.0.3
info:
  title: LiveKit Sarvam Voice Agent API
  description: |
    REST API for managing voice agents, knowledge bases, sessions, and analytics.
    
    ## Features
    - **Voice Agent Management** - Create and manage AI voice agents
    - **Knowledge Base** - Upload documents and search with RAG
    - **Session Management** - Track voice call sessions
    - **Analytics** - Monitor performance and business metrics
    - **Multi-tenant** - Organization-based isolation
    
    ## Authentication
    Currently using organization/tenant IDs for access control.
    Future versions will include API key authentication.
    
    ## Base URL
    - Development: `http://localhost:8000`
    - Production: `https://your-domain.com`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      description: Get all organizations with pagination
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Maximum number of results
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  total:
                    type: integer
                    example: 42
...
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: List agents
      description: Get all agents for an organization
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
          description: Organization ID or slug
          example: acme-corp
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
...
  /api/v1/sessions:
    get:
      tags: [Sessions]
      summary: List sessions
      description: Get sessions for an organization
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
...
  /api/v1/calls:
    get:
      tags: [Calls]
      summary: List calls
      description: Get call history
      parameters:
        - name: organization_id
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of calls
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items:
                      $ref: '#/components/schemas/CallSummary'
...
  /api/v1/analytics:
    get:
      tags: [Analytics]
      summary: Today's analytics overview
      description: Get comprehensive statistics for today
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Today's statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayStats'
  /api/v1/analytics/sessions:
    get:
      tags: [Analytics]
      summary: Recent sessions
      description: Get recent call sessions with details
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of recent sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSummary'
                  total:
                    type: integer
...
  /api/v1/analytics/agents:
    get:
      tags: [Analytics]
      summary: All agents performance
      description: Compare performance across all agents
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Agent performance comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPerformance'
components:
...
  schemas:
    Organization:
      type: object
      properties:
        _id:
          type: string
          example: j57abc123def
        slug:
          type: string
          example: acme-corp
        name:
          type: string
          example: Acme Corporation
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: number
          description: Unix timestamp
    Agent:
      type: object
      properties:
        _id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        role:
          type: string
        systemPrompt:
          type: string
        config:
          type: object
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: number
        updatedAt:
          type: number
    Session:
      type: object
      properties:
        _id:
          type: string
        sessionId:
          type: string
        organizationId:
          type: string
        agentId:
          type: string
        roomName:
          type: string
        participantIdentity:
          type: string
        callType:
          type: string
          enum: [inbound, outbound, web]
        status:
          type: string
          enum: [active, completed, failed, expired]
        startedAt:
          type: number
        endedAt:
          type: number
        durationSeconds:
          type: number
        metadata:
          type: object
    CallSummary:
      type: object
      properties:
        session_id:
          type: string
        agent_id:
          type: string
        status:
          type: string
        started_at:
          type: number
        duration_seconds:
          type: number
        call_type:
          type: string
...
    TodayStats:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            today:
              type: object
              properties:
                total_calls:
                  type: integer
                completed_calls:
                  type: integer
                failed_calls:
                  type: integer
                active_calls:
                  type: integer
                completion_rate:
                  type: integer
                avg_duration_seconds:
                  type: number
                total_duration_seconds:
                  type: number
            timestamp:
              type: string
    SessionSummary:
      type: object
      properties:
        session_id:
          type: string
        agent_id:
          type: string
        status:
          type: string
        started_at:
          type: number
        ended_at:
          type: number
        duration_seconds:
          type: number
        call_type:
          type: string
...
    AgentPerformance:
      type: object
      properties:
        status:
          type: string
        agents:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: string
              agent_name:
                type: string
              total_calls:
                type: integer
              completed_calls:
                type: integer
              failed_calls:
                type: integer
              completion_rate:
                type: integer
              avg_duration_seconds:
                type: number
              total_duration_seconds:
                type: number
              last_call_at:
                type: number
              status:
                type: string
        total_agents:
          type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Future authentication mechanism

# Future authentication
# security:
#   - ApiKeyAuth: []